refactor : 간단한 리팩토링
